Creating a cookie session
Add support for cookie sessions in requests

This commit adds functionality to the `urlinfo.py` script to support the use of cookie sessions when making requests to a URL. If the user has a cookie session, they can enter it at the prompt when running the script. The script then adds the session to the HTTP headers of the request using the `Cookie` field. If the user does not have a cookie session, the script sends the request without any headers.

The new functionality improves the user experience by allowing them to use the script with websites that require a cookie session for access. The code has been tested and works as expected.